+++ src/main/java/com/azure/app/BlobImageProvider.java
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
 package com.azure.app;
 
 import com.azure.core.http.policy.HttpLogDetailLevel;
