+++ src/main/java/com/azure/app/LocalDocumentProvider.java
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
 package com.azure.app;
 
 import org.apache.commons.io.FilenameUtils;
