+++ src/main/java/com/azure/app/LocalImageProvider.java
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
 package com.azure.app;
 
 import org.apache.commons.io.FilenameUtils;
@@ -152,9 +155,8 @@ public class LocalImageProvider implements BookCollector.ImageProvider {
 
     @Override
     public Mono<Void> deleteImage(Book book) {
-        return Mono.just(Paths.get(System.getProperty("user.dir"),
-            book.getCover().getPath()).toFile().delete()).map(result -> {
-            if (result.booleanValue() == false) {
+        return Mono.just(Paths.get(System.getProperty("user.dir"), book.getCover().getPath()).toFile().delete()).map(result -> {
+            if (!result) {
                 return Mono.error(new IllegalStateException("Image wasn't deleted."));
             } else {
                 deleteEmptyDirectories();
